# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: OTMNow Delivery Command Center API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /resources:
    get:
      operationId: listResources
      tags: [resources]
      summary: List resources/consultants
      parameters:
        - name: role
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: skill
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of resources
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Resource"
    post:
      operationId: createResource
      tags: [resources]
      summary: Create a resource
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateResourceBody"
      responses:
        "201":
          description: Created resource
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resource"
  /resources/{id}:
    get:
      operationId: getResource
      tags: [resources]
      summary: Get a resource
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Resource
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resource"
...
components:
  schemas:
...
    Resource:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        title:
          type: string
        practiceArea:
          type: string
          enum: [implementation, cloud_migration, ams, qa_certification, integration, development, data_migration, rate_maintenance, pre_sales]
        skills:
          type: array
          items:
            type: string
        utilizationTarget:
          type: integer
        currentUtilization:
          type: integer
        status:
          type: string
          enum: [available, allocated, over_allocated, bench, soft_booked, on_leave]
        hourlyRate:
          type: number
        location:
          type: string
        isContractor:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - name
        - practiceArea
        - status
        - createdAt
    CreateResourceBody:
      type: object
      properties:
        userId:
          type: integer
        name:
          type: string
        title:
          type: string
        practiceArea:
          type: string
        skills:
          type: array
          items:
            type: string
        utilizationTarget:
          type: integer
        hourlyRate:
          type: number
        location:
          type: string
        isContractor:
          type: boolean
        status:
          type: string
      required:
        - name
        - practiceArea
...